Refactor event handlers
